requirements.txt
# Justice AI — Python dependencies
# Tested with Python 3.11+

# ── Google ADK & AI ────────────────────────────────────────────────────────────
google-adk>=1.0.0
google-genai>=0.3.0
google-cloud-aiplatform>=1.40.0
google-auth>=2.25.0

# ── Web Framework ──────────────────────────────────────────────────────────────
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0
python-multipart>=0.0.9

# ── HTTP Clients ───────────────────────────────────────────────────────────────
httpx>=0.25.0
aiohttp>=3.9.0

# ── Data Processing ────────────────────────────────────────────────────────────
numpy>=1.24.0
pandas>=2.0.0

# ── Utilities ──────────────────────────────────────────────────────────────────
python-dotenv>=1.0.0
pyyaml>=6.0.0
setuptools>=65.0
wheel>=0.40.0
